fix: Refactor EditUserProfileCard and useEditUserProfileForm hooks
